reformat code structure for text module
