+++ neuralode/utils.py
@@ -1,8 +1,5 @@
 import numpy as np
 import torch
-from torch.utils.data import DataLoader
-import torchvision.datasets as datasets
-import torchvision.transforms as transforms
 
 
 def _check_none_zero(tensor, shape, device):
@@ -33,59 +30,3 @@ def runge_kutta(z0, t0, t1, f, eps=0.001):
         z = z + h / 6 * (k1 + 2 * k2 + 2 * k3 + k4)
         t = t + h
     return z
-
-
-def get_mnist_loaders(batch_size=128, test_batch_size=1000, perc=1.0):
-    transform_train = transforms.Compose([
-            transforms.RandomCrop(28, padding=4),
-            transforms.ToTensor(),
-            transforms.Normalize((0.5,), (0.5,)),
-        ])
-
-    transform_test = transforms.Compose([
-        transforms.ToTensor(),
-        transforms.Normalize((0.5,), (0.5,)),
-    ])
-
-    train_loader = DataLoader(
-        datasets.MNIST(root='.data/mnist', train=True, download=True, transform=transform_train), batch_size=batch_size,
-        shuffle=True, num_workers=2, drop_last=True
-    )
-
-    train_eval_loader = DataLoader(
-        datasets.MNIST(root='.data/mnist', train=True, download=True, transform=transform_test),
-        batch_size=test_batch_size, shuffle=True, num_workers=2, drop_last=True
-    )
-
-    test_loader = DataLoader(
-        datasets.MNIST(root='.data/mnist', train=False, download=True, transform=transform_test),
-        batch_size=test_batch_size, shuffle=False, num_workers=2, drop_last=True
-    )
-
-    return train_loader, test_loader, train_eval_loader
-
-
-def get_cifar_loaders(batch_size=128, test_batch_size=1000):
-
-    transform_train = transforms.Compose([
-        transforms.RandomCrop(32, padding=4),
-        transforms.RandomHorizontalFlip(),
-        transforms.ToTensor(),
-        transforms.Normalize((0.4914, 0.4822, 0.4465), (0.2023, 0.1994, 0.2010)),
-    ])
-
-    transform_test = transforms.Compose([
-        transforms.ToTensor(),
-        transforms.Normalize((0.4914, 0.4822, 0.4465), (0.2023, 0.1994, 0.2010)),
-    ])
-
-
-    train_loader = DataLoader(
-        datasets.CIFAR10(root='.data/cifar', train=True, download=True, transform=transform_train), batch_size=batch_size,
-        shuffle=True, num_workers=3)
-
-    test_loader = DataLoader(
-        datasets.CIFAR10(root='.data/cifar', train=False, download=True, transform=transform_test),
-        batch_size=test_batch_size, shuffle=False, num_workers=3)
-
-    return train_loader, test_loader, None
